% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_GOF.R
\name{AD}
\alias{AD}
\title{Run other GOF tests}
\usage{
AD(Z, Distr)
}
\arguments{
\item{Z}{a normalized random sample of n observations.}

\item{Distr}{a null distribution of the GOF test. It should be one of "Normal", "Logistic", "Cauchy", or "Gumbel}
}
\value{
A test statistic of the AD test.
}
\description{
Run the Anderson-Darling (AD) test.
}
\examples{

n=10
mu=2
sigma=1

X = rnorm(n, mu, sigma)

muhat = mean(X)
sighat = sd(X)

Z = (X-muhat)/sighat 
null_distr="Normal"
AD_teststat = AD(Z, null_distr)
AD_teststat
}
