% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BM_Library.R
\name{BM}
\alias{BM}
\title{Brownian Motion}
\usage{
BM(x = 0, t = 1, n = 10, d = 1)
}
\arguments{
\item{x}{a starting point of the Brownian motion, B(0).}

\item{t}{end of time, that is, \eqn{0=t_{0}<t_{1}<...<t_{n}=t}}

\item{n}{the last index of the time, that is, \eqn{t_{n}=t}}

\item{d}{a dimension of the Brownian motion: 1 or 2.}
}
\value{
A list of 
\describe{
\item{tn}{\eqn{t_{n}}}
\item{Bt1}{a vector of the Brownian motion (\eqn{B_{1}(t_{0},...,B_{1}(t_{n}))})}
\item{Bt2}{a vector of the Brownian motion (\eqn{B_{2}(t_{0},...,B_{2}(t_{n}))})}
}
}
\description{
Get a d-dimension Brownian motion B(t) and its graph
}
\examples{

#### Generate an 2-dimensional Brownian motion and its graph
x0=c(0,0)
lst = BM(x=x0, t=1, n=10, d=2)

}
