% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cauchy_table}
\alias{Cauchy_table}
\title{Examples of Cauchy tables}
\format{
A list containing examples of advanced calculated Cauchy tables:
\describe{
  \item{SMat}{Example of a valid input data which contains all required varialbe.}
  \item{VMat}{Dummy variable.}
  \item{xVec}{a partiton of [-1000,1000]}
  \item{ReVec}{Dummy variable.}
}
}
\usage{
data(Cauchy_table)
}
\description{
#####################
}
\examples{
data(Cauchy_table)
Cauchy_SMat = Cauchy_table$SMat
}
\keyword{datasets}
