% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_GOF.R
\name{CvM}
\alias{CvM}
\title{Run other GOF tests}
\usage{
CvM(Z, Distr)
}
\arguments{
\item{Z}{a normalized random sample of n observations.}

\item{Distr}{a null distribution of the GOF test. It should be one of "Normal", "Logistic", "Cauchy", or "Gumbel}
}
\value{
A vector of the CvM and Watson test statistics.
}
\description{
Run the Cramer-von Mises (CvM) and Watson tests.
}
\examples{

n=10
mu=2
sigma=1

X = rnorm(n, mu, sigma)

muhat = mean(X)
sighat = sd(X)

Z = (X-muhat)/sighat 
null_distr="Normal"
Vec = CvM(Z, null_distr)
CvM_teststat = Vec[1]
Watson_teststat = Vec[2]

CvM_teststat;Watson_teststat


}
