% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Distr_Info.R
\name{Distr_Information}
\alias{Distr_Information}
\title{Information of distribution}
\usage{
Distr_Information(x, strDistr)
}
\arguments{
\item{x}{a real number}

\item{strDistr}{a null distribution: "Normal", "Logistic", "Cauchy"}
}
\value{
a list of the following values:
\describe{
\item{fl}{f(x)}

\item{Fl}{F(x)}
\item{re}{r(x). Logistic only}
\item{Re}{R(x). Logistic only}
\item{phix}{\eqn{\phi(x)}}
\item{v0}{\eqn{v_{0}}}
\item{v1}{\eqn{v_{1}(x)}}
\item{v2}{\eqn{v_{2}(x)}}
\item{c0}{\eqn{c_{0}(x)}}
\item{c1}{\eqn{c_{1}(x)}}
\item{c2}{\eqn{c_{2}(x)}}
\item{Gamma}{\eqn{\Gamma(x)}}
\item{s1}{\eqn{s_{1}(x)}}
\item{s2}{\eqn{s_{2}(x)}}
\item{s3}{\eqn{s_{3}(x)}}
\item{S1}{\eqn{S_{1}(x)}}
\item{S2}{\eqn{S_{2}(x)}}
\item{S3}{\eqn{S_{3}(x)}}

}
}
\description{
Obtain all details regarding the null distribution such as c1(x).
}
\examples{
data(Normal_table)
x = 1.2
lst= Distr_Information(x, "Normal")
phix = lst$phix 
phix
c0 = lst$c0    
c0
}
