% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Graph_Obj.R
\name{DrawUnz}
\alias{DrawUnz}
\title{Implementing visual inspection of the KMT test statistic}
\usage{
DrawUnz(
  X,
  strDistr,
  type = "l",
  lty = 1,
  lwd = 1.5,
  col = "red",
  xlim = c(NA, NA),
  ylim = c(NA, NA),
  margin_x = 0.1,
  margin_y = 0.5,
  dGap = 0.01,
  b_abline = FALSE
)
}
\arguments{
\item{X}{a random sample of n observations}

\item{strDistr}{a null distribution for the hypothesis test: Normal, Cauchy, Logistic, or Gumbel.}

\item{type}{a type of  plot. A default is a line.}

\item{lty}{a line type. The default value is a solid.}

\item{lwd}{a line width.}

\item{col}{a line color.}

\item{xlim}{a limit for the x-axis.}

\item{ylim}{a limit for the y-axis.}

\item{margin_x}{a margin of graph in the x-axis.}

\item{margin_y}{a margin of graph in the y-axis.}

\item{dGap}{a length of subintervals of the x-axis.}

\item{b_abline}{a logical value for drawing vertical lines where the discontinuities of the graph happen.}
}
\value{
A list of the following values:
\describe{
\item{gObj}{plot of the graph of the supremand of the KMT test statistic, that is, |\eqn{U[n](z)}|  }
}
}
\description{
Draw the graph of |\eqn{U[n](z)}|.
}
\examples{
####################
n=20
mu0=2
sigma0=1
X = rnorm(n, mu0, sigma0)
strDistr="Normal"
DrawUnz(X, strDistr, type="l", lty=1, lwd=1.5, col="red",
                   xlim = c(-5,5), ylim = c(NA,NA),
                   margin_x=0.1, margin_y=0.5, dGap=0.01, b_abline=TRUE)

}
\references{
[1] Khmaladze, E.V., Koul, H.L. (2004). Martingale transforms goodness-of-fit tests in regression models. Ann. Statist., 32. 995-1034

[2] E.V. Khmaladze, H.L. Koul (2009). Goodness-of-fit problem for errors in nonparametric regression: distribution free approach. Ann. Statist., 37(6A) 3165-3185.

[3] Kim, Jiwoong (2020). Implementation of a goodness-of-fit test through Khmaladze martingale transformation. Comp. Stat., 35(4): 1993-2017
}
