% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Other_GOF.R
\name{KS}
\alias{KS}
\title{Run other GOF tests}
\usage{
KS(Z, Distr)
}
\arguments{
\item{Z}{a normalized random sample of n observations.}

\item{Distr}{a null distribution of the GOF test. It should be one of "Normal", "Logistic", "Cauchy", or "Gumbel}
}
\value{
A vector of the KS and Kuiper test statistics.
}
\description{
Run the Kolmogorov-Smirnov (KS) and Kuiper tests.
}
\examples{

n=10
mu=2
sigma=1

X = rnorm(n, mu, sigma)

muhat = mean(X)
sighat = sd(X)

Z = (X-muhat)/sighat 
null_distr="Normal"
Vec = KS(Z, null_distr)
KS_teststat = Vec[1]
Kuiper_teststat = Vec[2]

KS_teststat;Kuiper_teststat


}
