% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Logistic_table}
\alias{Logistic_table}
\title{Examples of Logistic tables}
\format{
A list containing examples of advanced calculated Logistic table:
\describe{
  \item{SMat}{Example of a valid input data which contains all required varialbe.}
  \item{VMat}{Dummy variable.}
  \item{xVec}{a partiton of [-30,13]}
  \item{ReVec}{a vector of R(x) for x in xVec.}
}
}
\usage{
data(Logistic_table)
}
\description{
#####################
}
\examples{
data(Logistic_table)
Logistic_ReVec = Logistic_table$ReVec
}
\keyword{datasets}
