% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Normal_table}
\alias{Normal_table}
\title{Examples of Normal tables}
\format{
A list containing examples of advanced calculated Normal tables:
\describe{
  \item{SMat}{Example of a valid input data which contains all required varialbe.}
  \item{VMat}{Dummy variable.}
  \item{xVec}{a partiton of [-5,5]}
  \item{ReVec}{Dummy variable.}
}
}
\usage{
data(Normal_table)
}
\description{
#####################
}
\examples{

data(Normal_table)
Normal_SMat = Normal_table$SMat
}
\keyword{datasets}
